+++ src/extensions/building/buildingext_hooks.cpp
@@ -2433,6 +2433,23 @@ DEFINE_HOOK(0x0042CAB9, _BuildingClass_Exit_Object_Factory_Busy_Customized_Alter
     return 0x0042CB16;
 }
 
+/**
+ *  Allows AI to repair Base Nodes by enabling the IsToRepair flag on those buildings.
+ *  Only applies in campaign, and only if the AIRepairBaseNodes under [AI] is set to yes/true. 
+ *
+ *  @author: JoyfulShush
+ */
+DEFINE_HOOK(0x0042A3D1, _BuildingClass_Unlimbo_AI_Repair_Base_Nodes, 5)
+{
+    GET(BuildingClass*, this_ptr, ESI);
+
+    if (Session.Type == GAME_NORMAL && !this_ptr->House->Is_Human_Player() && RuleExtension->AIRepairBaseNodes) {
+        this_ptr->IsToRepair = true;
+    }
+    
+    return 0;
+}
+
 
 /**
  *  Main function for patching the hooks.