+++ src/extensions/rules/rulesext.h
@@ -137,6 +137,11 @@ public:
     int AINavalYardAdjacency;
 
     /**
+     *  Should the AI automatically repair buildings built as Base Nodes?
+     */
+    bool AIRepairBaseNodes;
+
+    /**
      *  The "double penalty" or "half penalty". Multiply this by the power
      *  units you are short of to get the actual penalty to the build speed.
      */